requirements.txt
beautifulsoup4==4.12.3
numpy==1.24.3
opencv_python==4.8.0.76
Pillow==10.4.0
pyzbar==0.1.9
Requests==2.32.3
